Improves the styleguides for local components
For the RatioEnforcer I decided to remove the modifiers because they are
so specific they add little to the documentation. Maybe if they were
tied together with the images they are supposed to be used with, it
would be more helpful.

As for the Panels, I went to on the opposite direction, because I think
it is helpful to look at all the different panels we have (and the fact
they are visually pleasing doesn't hurt either).

I added some inline styles here and there, to make the styleguide look
nicer (IMO), I'm assuming whoever reads this will understand that the
inline styles are not to be copied, but maybe we should write that down
somewhere and assume nothing.