Page names should use dashes instead of underscores
It's much better for SEO, since dashes are treated as a space, whereas
underscores are a hard character

So a search for 'Design Sprint' can be matched to 'design-sprint', but not
'design_sprint'

I only changed this for actual page names, and not other files (images,
partials, etc). I'm not sure if we should keep file naming consistency
and change everything.