Fixes homepage SVGs
Why:

We were having weird problems with SVGs on Firefox. This could be
anything from a small typo in our code to a bug in Firefox itself. After
hours of debugging, we decided to move the SVGs to object elements,
which seems to work fine as well. There are extra requests being made
now, but at least it's working

This change addresses the need by:

* Moving all SVGs in the homepage to an `object` tag, instead of having them inline
* Refactor the JS code for the animations to interact with the contents of the object, instead of directly in the DOM