Redesign homepage and add OncoStats case study
This PR got out of hand very fast; it was supposed to be a small change,
but we were changing things so much that I decided to keep everything on the same branch until we were able to understand the redesign a little bit better. A lot of thins changed since we started it.

In this change, I'm redesigning the homepage and adding the OncoStats case study. Redesigning the homepage means also updating blue. The margins are different, the buttons are different, the text color is different. A lot of things changed.

The best part about all of this is that I got to remove a lot of code, which makes me happy. The website should be faster now, on webpagetest, it is taking 3s to first paint on the US. That's is too much.

Special remarks:

* Removing styleguide. We did remove styleguide, it was becoming a pain
to maintain and it wasn't providing any value.
* Removing newsletter. We no longer have a newsletter.