Add blog posts section to homepage
Why:

In this change, I'm adding a section with the most recent blog posts to
the homepage. Because there is no back-end for this website, we decided to load the blog posts dynamically from an endpoint I recently created on the blog.

You'll notice a new component named `SectionHeading`. It uses custom font-sizes because this component is intended for aesthetics, not content. The font-size works the same way a `height` would on any other component. For that reason, I didn't make an effort to include those sizes in the system.

This change addresses the need by:

* Add a section for blog posts.
* Load blog posts from API and use a template to display on the screen.