reduce our website by 1.8mb
In this change, I update the ventures' portfolio section to use a single
image for each venture. In the current implementation, we
load a color and a grey image for each venture. With this change, we
simply apply a css filter to grayscale the color image. This reduces the
size of the whole website in 1.8mb.