stop loading both mobile and desktop images
in some places, we use different images for both mobile and desktop.
it's not just a matter of the images having different resolutions, they
also have different "cuts", making them different images. because our
images are loaded async with JavaScript, the CSS rules that add
'display: none' to mobile images in desktop (and the other way around)
are ignored. this means that desktop and mobile images are always loaded
in any device.

to solve this, there's a thing called [art
direction](https://developer.mozilla.org/en-US/docs/Learn/HTML/Multimedia_and_embedding/Responsive_images#Art_direction)
that gatsby has support for.

extra: reduced the quality of hero-3 in desktop because it's the biggest
image in our website and it doesn't seem to change its final quality.