Optimize the imports of lodash (#226)
Importing lodash usign `import _ from "lodash"` will import the whole
library. The correct way to import lodash is to only import the pieces
you need, for instance `import map from lodash/map`. In this case, we
simply don't need lodash in this situation, so I removed it. It's still
being used in other places, but this already removes around 70kb to the
bundle size. And because this is JS and our initial animation
requires JS to run, we can use as many optimizations as we can.