Generating social media banners with ImageMagick
Why:

* Laura and Baila have been manually creating images for each post
  (e.g.: https://twitter.com/subvisual/status/1286280282354655234)
* That workflow probably doesn't produce the best tweets (see my previous post,
  just seconds ago), in terms of engagement

This PR adds a script that can be used to automatically generate
a banner similar to the one currently being generated manually.

It's not yet integrated into any CI workflow. It's a step towards that.

Also, the image generation itself is not yet finished. I need to tweak
font sizes and positioning